// frontend/src/api/model/recruitmentLinkType.ts
/**
 * Generated by orval v7.18.0 🍺
 * Do not edit manually.
 * Libre-Q API
 * OpenAPI spec version: 0.1.0
 */

/**
 * Enum for types of recruitment links.
 */
export type RecruitmentLinkType = (typeof RecruitmentLinkType)[keyof typeof RecruitmentLinkType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const RecruitmentLinkType = {
    public: 'public',
    individual: 'individual',
    limited: 'limited',
} as const;